Changed delete to backspace in line editor
In the line editor, both the delete and backspace keys were originally
implemented to behave as expected. However, It turns out that many
terminal emulators actually send a delete character when the backspace
key is pressed. Since the backspace key is commonly used at the end of
the input (a location where the delete key would have no effect), many
people get confused and assume the key doesn't work at all.

As I don't have any reasonable way to tell what signals are going to be
generated by a terminal emulator at runtime, I decided the best solution
would just be to modify the behavior in the general case. Since the use
of backspace is much more common than the use of the delete key, I have
changed the behavior so that both the backspace and delete signals
perform a backspace. This way, the behavior is consistent, regardless of
the design of the terminal emulator.