Added a migration status screen
Previously, when a migration was run, the main menu screen would simply
display statically throughout the course of the run. This is
problematic, as it doesn't give the user any feedback concerning how far
the migration has come, how much is left to go, whether/where the
progress is frozen, and what (if any) failures occurred. Now, a status
screen with a live checklist, progress bar, and set of error counters is
displayed instead.