Improved checksum handling
It's possible that the nexus metadata files might not exist, or that
they might not contain required information, or that this information
might not be accurate. This update accounts for these possibilities.

Before, the tool assumed a .nexus/attributes file existed for every
artifact, and used the repo path and checksums from that file in its
calculations. It also assumed that a .md5 and .sha1 file both existed,
and deployed both of these after each artifact. Now, it first checks for
a .nexus/attributes file, then a .md5 and .sha1 file, and finally it
hashes the file manually, so hashes will always be available. For the
repo path, it first checks the .nexus/attributes file, and then it uses
the filesystem path.

Under the repository configuration menu, there is now a new option. If
this option is enabled, none of the Nexus metadata will be trusted
during migration. This means all repo paths are generated via the
filesystem path, and all checksums are generated by hashing the
artifacts themselves. This is rather slow, but is guarenteed to give
correct metadata, and so is useful if you believe the Nexus metadata is
incorrect or corrupted.