Fixed issue with virtual repo migration
We make one request to Artifactory to see what repositories exist. Then,
for each repository we're migrating, we use POST if it's in that
list (for update) and PUT if it's not (for create). However, empty
virtual repositories don't appear to show up in the list, and using PUT
with an existing repository just 409s. Therefore, we now check each
virtual repository before sending the POST or PUT, just to be sure if it
actually exists.