Fixed file read issue
In some cases, we might see a file on the filesystem, and then get a
FileNotFound exception when trying to read it (for example, if the file
is a softlink to a nonexistant file). There was no handling for this, so
a file read thread would just exit immediately on this sort of error. If
all four threads ran into these issues, the migration would either end
early or deadlock. Now, we handle these errors appropriately.