Added LDAP password extraction
The LDAP password is stored in the nexus config file, but it's
obfuscated, and therefore useless in its available form. Therefore, you
would need to write in the LDAP password into the migrator to make LDAP
work. Now though, the migrator automatically decrypts the password from
the config file, so it doesn't need to be written in manually anymore.