Major refactor
Refactored the codebase to increase separation of concerns and reduce
spaghetti hacks. Editing the codebase should be much less of an ordeal
now.

If the logs are set to 'debug', the tool logs every keypress. This could
be a problem when used with a password field, as the password would
appear in plaintext in the logs. Now, logged password fields are
automatically masked with asterisk characters.