Fixed some small issues
During the artifact and post migration steps, some of the text was
wrong. Now it displays the correct information.

Passing --silent is supposed to suppress stdout logging in
non-interactive mode, but it was also suppressing file logging. Now,
passing --silent only silences stdout, but not logs to a file.

In Windows, if there wasn't a logger configured, some logs would be
printed to the UI by PDCurses, which could potentially break the
interface. Now, if no other logger is configured, a default "nil" logger
with no behavior is used, which circumvents this issue.