Added basic repo target conversion
Before, the tool could not convert from Nexus repo targets to
Artifactory permission patterns. It would check for one of the three
default regex patterns, and convert those, but any other regexes would
require manual conversion by the user.

Now, logic for automatically parsing and converting Nexus regex patterns
to Artifactory Ant pattern sets has been added. This is subject to the
following restrictions:

Only Nexus 2 repo targets are converted at the moment. Support for
Nexus 3 content selectors will be added in the future.

Many regex features are not yet supported. The following features will
be supported in the future:
- \p{...} and \P{...} character classes
- backreferences (\1, \2, etc)
- word and text boundaries (\b, \B, ^, $, etc)
- lookarounds (positive/negative lookaheads/lookbehinds)

Regexes containing an unbounded quantifier (*, +, {n,}, etc) applied to
anything other than a . or [^/] character set will not be converted.
This is because the only unbounded quantifiers in Ant patterns
are * ([^/]*) and ** (.* roughly). Therefore, patterns like \d* or (x.)+
are impossible to convert, and the user should manually find a suitable
replacement.

Some types of regexes can technically be converted, but will end up
converting to hundreds or thousands of Ant patterns. This has the
potential to substantially slow down both the migrator and Artifactory,
and can be impossible to maintain. Often, the correct action in this
case is to manually find a suitable replacement pattern instead.
Therefore, if the number of patterns generated is over a certain
threshold (10 by default), the conversion will not happen.