Fixed two artifact upload bugs
Fixed a bug that caused the tool to only do special repo type handling
on repos who's names have not been changed.

Fixed a bug that caused the tool to attempt to migrate all artifacts
from Nexus 3, even when the repositories the artifacts belong to are not
marked for migration.