Fixed multiple bugs
When some parts of the migration failed, the entire migration ended,
even where it would make sense to just log an error and move on. Now,
more errors are caught and logged properly so as to not interrupt the
migration as a whole.

When there were errors during migration, the tool would still say
"Migration successful!", because the migration wasn't interrupted. This
is confusing and unintuitive. Now, if the migration completes with
errors, the tool says "Migration successful, but errors occurred."

Artifactory only allows LDAP settings to have names that are valid
non-namespaced XML names, but the tool allowed any non-empty name. Now
it restricts LDAP names properly.