Fixed Nexus 3 role migration bug (#25)
When creating a repository in Nexus 3, some default view and admin
privileges for the repository are generated automatically. When deleting
a privilege, usually the privilege is removed from any roles it belongs
to. However, when deleting a repository, the default privileges created
for that repository are deleted, but they are not removed from roles.
This appears to be a bug in Nexus 3, and causes the migrator to
encounter an error when reading roles.

Now, this error is circumvented (instead, the privilege entry in the
role is simply ignored) if a missing auto-created privilege is
encountered in a role.