Added logging for generating submenus
All the big menus get debug logs when generated, but their submenus (of
which there may be many) do not. This can make it difficult to tell why
opening a menu might take a long time, and whether the submenus are
being processed at all. Now, each time a submenu is processed, its name
is printed to the debug log.