Made SSL fix optional
A while back, a workaround was added for an SSL bug that forced the tool
to use TLSv1.0, rather than a later version. Now that TLSv1.0 is
deprecated, this workaround no longer makes sense. It has been turned
off by default, and can be enabled by passing the -o option.