Moved log summary to CRITICAL so it always shows
Fixes #34. The summary log in non-interactive mode used to be logged on
the INFO level. This is a problem because you can't restrict the log
level to something above INFO (WARN or ERROR) without hiding the
summary. Now, the summary is logged on the CRITICAL level, which means
it will always be logged.